Declaration of input variables.
size_image - image sizes are copied.
n_size - The user enters the size of the window around each pixel.
k_factor - The user enters the filter efficiency in percent.
border - the thickness of the border around the pixel is calculated.
k & l - coefficients from the formula.
Then the condition of the formula is fulfilled.
At the end, unnecessary variables are removed.